+++ examples/custom_frontend/app.py
@@ -3,13 +3,13 @@ from typing import List
 from pinferencia import Server
 
 
-def stat(data: List[float]) -> dict:
-    return {
-        "mean": sum(data) / len(data),
-        "max": max(data),
-        "min": min(data),
-    }
+def stat(data: List[float]) -> float:
+    return sum(data)
 
 
 service = Server()
-service.register(model_name="stat", model=stat, metadata={"task": "Stat"})
+service.register(
+    model_name="stat",
+    model=stat,
+    metadata={"display_name": "Awesome Model"},
+)
